task_2.3.1_v3

+++ src/main/java/org/example/crud/dao/UserDaoImpl.java
@@ -21,7 +21,7 @@ public class UserDaoImpl implements UserDao{
     }
 
     @Override
-    public User getUserById(int id) {
+    public User getUserById(long id) {
        return entityManager.find(User.class, id);
     }
